+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.1.5 (October 16, 2018)
+- Add `overrideHttpPut` option for handling environments not allowing http put.
+
 ## v0.1.4 (June 21, 2018)
 - Fix superagent `attachCookies` and `saveCookies` usage